implement month remembering
