+++ frontend/src/app/components/routing/wp-calendar/wp.calendar.component.html
@@ -23,13 +23,6 @@
 
     <filter-container></filter-container>
 
-    <accessible-by-keyboard
-            linkClass="hidden-for-sighted skip-navigation-link"
-            [linkAriaLabel]="text.text_jump_to_pagination"
-            (execute)="setAnchorToNextElement()">
-        <span [textContent]="text.jump_to_pagination"></span>
-    </accessible-by-keyboard>
-
     <wp-calendar [projectIdentifier]="projectIdentifier">
     </wp-calendar>
 </div>
