+++ frontend/src/app/components/wp-calendar/wp-calendar.component.ts
@@ -9,28 +9,32 @@ import {WorkPackageCollectionResource} from "core-app/modules/hal/resources/wp-c
 import {WorkPackageTableFiltersService} from "core-components/wp-fast-table/state/wp-table-filters.service";
 import {Moment} from "moment";
 import {WorkPackagesListService} from "core-components/wp-list/wp-list.service";
+import {StateService} from "@uirouter/core";
+import {UrlParamsHelperService} from "core-components/wp-query/url-params-helper";
+import * as moment from "moment";
 
 @Component({
   templateUrl: './wp-calendar.template.html',
   selector: 'wp-calendar',
 })
 
-export class WorkPackagesCalendarController implements OnInit, OnDestroy, AfterViewInit {
+export class WorkPackagesCalendarController implements OnInit, OnDestroy {
   calendarOptions:Options;
   events:any;
   @ViewChild(CalendarComponent) ucCalendar:CalendarComponent;
   @Input() projectIdentifier:string;
 
   constructor(readonly states:States,
+              readonly $state:StateService,
               readonly wpTableFilters:WorkPackageTableFiltersService,
               readonly wpListService:WorkPackagesListService,
-              readonly tableState:TableState) {
+              readonly tableState:TableState,
+              readonly urlParamsHelper:UrlParamsHelperService) {
   }
 
   ngOnInit() {
     // Clear any old table subscribers
     this.tableState.stopAllSubscriptions.next();
-    let calendar = this.ucCalendar;
 
     this.calendarOptions = {
       editable: false,
@@ -71,24 +75,35 @@ export class WorkPackagesCalendarController implements OnInit, OnDestroy, AfterV
     });
   }
 
-  ngAfterViewInit() {
-    //setTimeout(() => this.ucCalendar.fullCalendar('gotoDate', '2018-09-01'));
-  }
-
   ngOnDestroy() {
     // nothing to do
   }
 
   public updateTimeframe($event:any) {
-    let calendarView = jQuery($event.currentTarget).fullCalendar('getView')!;
+    let calendar = jQuery($event.currentTarget);
+    let calendarView = calendar.fullCalendar('getView')!;
     let startDate = (calendarView.start as Moment).format('YYYY-MM-DD');
     let endDate = (calendarView.end as Moment).format('YYYY-MM-DD');
 
-    if (!this.wpTableFilters.currentState) {
+    if (!this.wpTableFilters.currentState && this.states.query.resource.value) {
+      // nothing to do
+    } else if (!this.wpTableFilters.currentState && this.$state.params.query_props) {
+      let query_props = this.$state.params.query_props;
+      let hash_props = JSON.parse(decodeURIComponent(this.$state.params.query_props));
+
+      let datesIntervalFilter = _.find(hash_props.f, { 'n': 'datesInterval'})!;
+      let lower = moment(datesIntervalFilter.v[0]);
+      let upper = moment(datesIntervalFilter.v[1]);
+
+      this.wpListService.fromQueryParams({ query_props: query_props }, this.projectIdentifier).toPromise().then(() => {
+        calendar.fullCalendar('gotoDate', lower.add(upper.diff(lower, 'days') / 2, 'days').format('YYYY-MM-DD'));
+      });
+    } else if (!this.wpTableFilters.currentState) {
       let query_props = `{%22f%22:[{%20%22n%22:%20%22status%22,%20%22o%22:%20%22o%22,%20%22v%22:[]%20},%20{%20%22n%22:%22datesInterval%22,%20%22o%22:%20%22%3C%3Ed%22,%20%22v%22:%20[%22${startDate}%22,%20%22${endDate}%22]%20}],%22pp%22:50}`;
 
       this.wpListService.fromQueryParams({ query_props: query_props }, this.projectIdentifier).toPromise();
     } else {
+      let params = this.$state.params;
       let filtersState = this.wpTableFilters.currentState;
 
       let datesIntervalFilter = _.find(filtersState.current, {'id': 'datesInterval'})!;
